fix for connection issue after rename
